# CSADigitizer: last threshold crossing for ToT