--- a/etc/unittests/test_modules/test_06-11_digitization_csa_last_threshold_crossing_tot.conf
+++ b/etc/unittests/test_modules/test_06-11_digitization_csa_last_threshold_crossing_tot.conf
@@ -1,0 +1,33 @@
+[Allpix]
+detectors_file = "detector.conf"
+number_of_events = 1
+random_seed = 0
+
+[DepositionPointCharge]
+model = "fixed"
+source_type = "point"
+position = 445um 220um 0um
+number_of_charges = 1000
+
+[ElectricFieldReader]
+model = "linear"
+bias_voltage = 100V
+depletion_voltage = 150V
+
+[GenericPropagation]
+temperature = 293K
+charge_per_step = 100
+propagate_electrons = false
+propagate_holes = true
+
+[PulseTransfer]
+
+[CSADigitizer]
+log_level = DEBUG
+model = "simple"
+clock_bin_toa = 1ns
+clock_bin_tot = 5ns
+sigma_noise = 5mV
+last_threshold_crossing_tot = true
+
+#PASS [R:CSADigitizer:mydetector] Pixel (2,0): time 13clk, signal 14clk