This PR aims to add an option to allow to use the last threshold crossing in the ToT measurement to determine the ToT. This is the behavior on the MuPix10 and allows better precision in high noise to signal scenarios.

To demonstrate the difference, this plot was made in a high noise scenario (Fe55) with a MuPix10 parameterization:

![new_tot_mode](https://user-images.githubusercontent.com/36662302/115856139-65e5b000-a42c-11eb-8b0e-11a53ea85985.png)

The current implementation has two problems:
- no option to enable or disable the mode right now
- currently this is bound to the integration time, which is bad since we ideally want this to be a certain amount of time after the ToA, i.e. some max amounts of clock cycles where it stops

Regarding the second one:
I think it would be best to introduce two integration times: a ToA integration time, and a ToT integration time. First, we amplify the pulse only for the ToA integration time. If we get a signal, then we amplify again so that we end up at the ToT integration time.

The simpler alternative would be to just define a ToT integration time, and stop when `time - toa_time > tot_time`. The problem here is that we can't ensure that the pulse is amplified long enough (imagine the integration time is 1us and the ToT integration time 100ns, but the ToA time is 950ns). To prevent this we would need a relatively high base integration time.

Regarding the first one:
This behavior, no matter which option, should definitely be optional as is it costs some performance, and also maybe other chips don't have this behavior. I haven't run benchmarks (yet), but it felt not insignificant. With the first option, I propose this interface:
```config
tot_last_crossing = true
toa_integration_time = 1us
tot_integration_time = 1us
```
If `tot_last_crossing` is false, the interface would stay just as before.

For the second option, one can simply use
```config
last_crossing_integration_time = 1us
```
If this variable is set, the new mode is enabled, else it is not.


Any thoughts / feedback?