@stephanlachnit 2021-04-23
Performance numbers (10,000 runs):

New behavior:
```
(S) Executed 7 instantiations in 43 seconds, spending 89% of time in slowest instantiation CSADigitizer:detector
(I)  Module GeometryBuilderGeant4 took 0.0727533 seconds
(I)  Module DepositionGeant4 took 15.2367 seconds
(I)  Module ElectricFieldReader:detector took 0.0654382 seconds
(I)  Module GenericPropagation:detector took 11.5512 seconds
(I)  Module PulseTransfer:detector took 2.43689 seconds
(I)  Module CSADigitizer:detector took 237.633 seconds
(I)  Module ROOTObjectWriter took 0.552364 seconds
(S) Average processing time is 4 ms/event, event generation at 232 Hz
(S) This corresponds to a processing time of 32 ms/event per worker
```
Old behavior:
```
(S) Executed 7 instantiations in 38 seconds, spending 86% of time in slowest instantiation CSADigitizer:detector
(I)  Module GeometryBuilderGeant4 took 0.0656114 seconds
(I)  Module DepositionGeant4 took 13.2465 seconds
(I)  Module ElectricFieldReader:detector took 0.0303207 seconds
(I)  Module GenericPropagation:detector took 15.4272 seconds
(I)  Module PulseTransfer:detector took 2.29964 seconds
(I)  Module CSADigitizer:detector took 192.944 seconds
(I)  Module ROOTObjectWriter took 0.609226 seconds
(S) Average processing time is 4 ms/event, event generation at 266 Hz
(S) This corresponds to a processing time of 32 ms/event per worker
```

So the new behavior takes ~25% longer. Obviously this depends on the settings: with high noise, the earlier implementation sets the ToT too early, which wouldn't happen in lower noise cases. Also, the integration time was not "optimized" (3us, but the ToT never goes higher than 800ms). With the first implementation, one also saves some computing time by only amplifying for the ToA time if the signal doesn't cross the threshold (only 0.04% of the events create a hit).