@stephanlachnit 2021-04-23
So I played around with the first idea of the separate ToA and ToT integration times, it sounds nice at first but as far as I see it the pulse needs to be amplified from the beginning again if we cross the threshold, so it is potentially even worse than just having a longer base integration time.