@stephanlachnit 2021-04-24
> Is what you are describing here something similar to what is usually referred to "iToT" or integrated ToT? Summing up the different sections above threshold? Or really just returning the maximum ToT found in the event?

Yes, afaik it is the last crossing.

> I think it is interesting but we should be a bit careful not to add too many "small" features to the module - because every front-end in the end might require a slight tuning here and there...

Yes I can understand that. The MuPix10 does have some other features that wouldn't fit in here, for example a second threshold. I'm not sure if I will ever implemented everything, but just out of interest: if I would fork the module, would you be interested in having this upstream?