@stephanlachnit 2021-04-24
> > Yes I can understand that. The MuPix10 does have some other features that wouldn't fit in here, for example a second threshold.
> 
> Keep in mind that also the modularity of APSQ can provide you with some features - e.g. setting up two digitizer modules that process the same incoming PixelCharge objects with different thresholds and write the output to trees for a comparison/merging offline is done with two lines of config file:
> 
> ```toml
> [CSADigitizer]
> output = "threshold1"
> threshold = 800e
> 
> [CSADigitizer]
> output = "threshold2"
> threshold = 1200e
> ```

Ah thanks, I didn't know that.

> > I'm not sure if I will ever implemented everything, but just out of interest: if I would fork the module, would you be interested in having this upstream?
> 
> You mean a fully-specialized module for MuPix10? I haven't wrapped my head around this yet - let's see once we get there...

Yes, just in theory. E.g. the MuPix10 also has a ramp mode, which changes the threshold linearly after TS1 is latched, which would definitely not fit in this module (but it also is something that I don't need right now and probably exceeds the scope of my bachelor thesis). And another thing is that a colleague mentioned that the falling slope of the amplified pulse is independent of the charge, which also would require major rewrites of the module that are probably better suited in a separate one.