@stephanlachnit 2021-04-26
@simonspa The code is ready to review, I'm not entirely happy with the name `last_threshold_crossing_tot`.
I *think* the last commits is a bit more efficient by only assigning the `was_above_threshold` a new (hard-coded) value when it actually changes, but I'm not entirely sure and also it's harder to read.